Basic Calculator
A simple calculator, doesn't have buttons, rather takes two numbers and the operator typed in it.
In the current version a button needs to be clicked to reuse the previous value.
The next update intends to fix it by removing the button and assigning the work to a keyboard shortcut instead.